make intro notebook runnable without installing document converters
