+++ tapeagents/examples/data_science.py
@@ -5,17 +5,17 @@ import tempfile
 
 from tapeagents.agent import Agent
 from tapeagents.core import Action, AgentEvent, FinalStep, Observation, Tape
-from tapeagents.environment import Environment
+from tapeagents.environment import CodeExecutionEnvironment, Environment
 from tapeagents.utils import run_in_tmp_dir_to_make_test_data
 from tapeagents.view import Call, Respond
-from tapeagents.tools.container_executor import ContainerExecutor
+from tapeagents.container_executor import ContainerExecutor
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
 
 import sys
 
 from tapeagents.autogen_prompts import AUTOGEN_ASSISTANT_SYSTEM_MESSAGE
-from tapeagents.collective import CollectiveAgent, CollectiveEnvironment, CollectiveTape
+from tapeagents.collective import CollectiveAgent, CollectiveTape
 from tapeagents.develop import Develop
 from tapeagents.llms import LLM, LiteLLM
 from tapeagents.rendering import BasicRenderer, PrettyRenderer
@@ -52,7 +52,7 @@ def make_world(llm: LLM | None = None) -> tuple[Agent, Tape, Environment]:
     )
     start_tape = CollectiveTape(context=None, steps=[])
     now = f"{datetime.datetime.now():%Y%m%d%H%M%S}"
-    env = CollectiveEnvironment(ContainerExecutor(work_dir=f"outputs/data_science/{now}"))
+    env = CodeExecutionEnvironment(ContainerExecutor(work_dir=f"outputs/data_science/{now}"))
     return org, start_tape, env
